requirements.txt
aiohttp
python-socketio
websocket